ai/requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
ultralytics==8.0.3
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
Easyocr

# Logging -------------------------------------
tensorboard>=2.4.1


# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Web
flask
flask_cors
